refactor(fe): 챗봇 회사 전역 floating button으로 ui 변경
